+++ e2e/medications-wizard-biweekly.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "@playwright/test";
+import { expect, test } from "./setup/test";
 
 import {
   STUB_MEDICATION_ID,